chore: MA-29 Destination 정리
