+++ app/src/main/java/com/mundcode/muntam/navigation/Destination.kt
@@ -1,0 +1,5 @@
+package com.mundcode.muntam.navigation
+
+interface Destination {
+    val route: String
+}
